Upload Debug
It appears the latest version of GoldRaccoon isn’t uploading the full
files that are created… This might be an issue with the Xcode beta but
it’s more likely a problem on GoldRaccoon’s side. TODO a bit of
debugging on the latest changes to the GoldRaccoon API - might end up
rolling it back.